Change to use Google App Engine and ndb for storage
This code is very similar to the original Flask + SQLAlchemy code. Instead of the constructor for Memory, we have a create() classmethod that serves the same purpose.
ndb.Model already has a good __repr__, so I removed that. I left the to_dict, even though Jinja is more than capable of pulling items out of the object.

I replaced double quotes with single (because of my own personal preferences).

I'm not sure why in the danbooru command it uses %20 instead of + to encode spaces in the query string. I use urllib.quote_plus to do this more fully
(in case people put other strange characters in the search string).

Switched use of requests to urlfetch, which is the de facto method of getting remote resources via HTTP.

 event() and stream() are currently commented because I can't use gevent here and haven't figured out which method of SSE/pubsub/polling I want to use.